Added maps to the file structure for organisation
