+++ Halfway/Components/Triangle.swift
@@ -1,0 +1,22 @@
+//
+//  Triangle.swift
+//  Halfway
+//
+//  Created by Johannes on 2020-10-10.
+//  Copyright © 2020 Halfway. All rights reserved.
+//
+
+import SwiftUI
+
+struct Triangle: Shape {
+    func path(in rect: CGRect) -> Path {
+        var path = Path()
+
+        path.move(to: CGPoint(x: rect.midX, y: rect.minY))
+        path.addLine(to: CGPoint(x: rect.minX, y: rect.maxY))
+        path.addLine(to: CGPoint(x: rect.maxX, y: rect.maxY))
+        path.addLine(to: CGPoint(x: rect.midX, y: rect.minY))
+
+        return path
+    }
+}
