+++ Halfway/Extensions/ViewToUIImage.swift
@@ -10,20 +10,22 @@ import SwiftUI
 import UIKit
 
 extension View {
-// This function changes our View to UIView, then calls another function
-// to convert the newly-made UIView to a UIImage.
+// This function changes SwiftUI View to UIView, then calls another function
+// to convert that UIView to a UIImage.
     public func asUIImage() -> UIImage {
         let controller = UIHostingController(rootView: self)
         
         controller.view.frame = CGRect(x: 0, y: CGFloat(Int.max), width: 1, height: 1)
+        
+        //Making the background transparent in case the View is not rectagular
         controller.view.backgroundColor = UIColor.white.withAlphaComponent(0)
+        
         UIApplication.shared.windows.first!.rootViewController?.view.addSubview(controller.view)
         
         let size = controller.sizeThatFits(in: UIScreen.main.bounds.size)
         controller.view.bounds = CGRect(origin: .zero, size: size)
         controller.view.sizeToFit()
         
-// here is the call to the function that converts UIView to UIImage: `.asImage()`
         let image = controller.view.asUIImage()
         
         controller.view.removeFromSuperview()
@@ -32,7 +34,7 @@ extension View {
 }
 
 extension UIView {
-// This is the function to convert UIView to UIImage
+// Converts UIView to UIImage
     public func asUIImage() -> UIImage {
         let renderer = UIGraphicsImageRenderer(bounds: bounds)
         return renderer.image { rendererContext in
